Cleaned java code (part 2).

+++ tests/gdx-tests/src/com/badlogic/gdx/tests/GLProfilerErrorTest.java
@@ -1,12 +1,9 @@
 /*******************************************************************************
  * Copyright 2015 See AUTHORS file.
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
  *   http://www.apache.org/licenses/LICENSE-2.0
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -76,7 +73,7 @@ public class GLProfilerErrorTest extends GdxTest {
 
 		int x = 10;
 		int y = Gdx.graphics.getHeight() - 10;
-		y -= font.draw(batch, "e - Enable debugging\n" + "d - Disable debugging\n" + "l - Test log error listener\n"
+		y -= (int)font.draw(batch, "e - Enable debugging\n" + "d - Disable debugging\n" + "l - Test log error listener\n"
 			+ "t - Test throw error listener\n" + "c - Test custom listener\n\n" + "Expected error: GL_INVALID_VALUE (0x501, 1281)",
 			x, y).height;
 		y -= 10;
